* Added Data reader for Speech.mul. * AsciiSpeechPacket filename fixed. Packet can now send keywords from Speech.mul. You can now pick up and drop items in your inventory and other containers! * Support in UILegacy/Cursor.cs for holding item. * UltimaClient now correctly sets item parents when placing an item in a container. * Fixed a bug with all containers sharing a single static contents list. * Added container contents checking so items are not added multiple times. * Closing the WorldScene now properly resets the UI. * Removed unused EventSystem code. * Added new Interaction class, which will be used to interact with the server. Interaction currently supports SendChat, PickUpItem, and DropItem. More to come. * Added routines to Item and Container to facilitate Picking Up and Disposing actions. * Fixed a bug where removing a worn item would recursively call its own Dispose() routine. * Hitting Enter while typing text in a TextEntry gumpling will activate the owning Gump. * ContainerGump update which only deletes/adds item textures as necessary. * ItemGumplings now dispose themselves when their item is Disposed.

+++ UltimaXNA/UltimaXNA/Network/Packets/Client/PickupItemPacket.cs
@@ -26,7 +26,7 @@ namespace UltimaXNA.Network.Packets.Client
 {
     public class PickupItemPacket : SendPacket
     {
-        public PickupItemPacket(Serial serial, short count)
+        public PickupItemPacket(Serial serial, int count)
             : base(0x07, "Pickup Item", 7)
         {
             Stream.Write(serial);
